Changed Video ID
Changed the output directory of videos so each video is put in a new
output directory. May change this in the future, but this is needed
since it is possible to have 2 videos with the same name. The video
names are created based on the current time, but with multithreading it
is possible to have 2 videos packages at the exact same time.